fix(cli): rename moves raw pending files; test file split under 400 LOC
Codex P2 x2: (1) an interrupted explicit file pairing leaves a
pending credential FILE before the machine-wide marker exists — the
routed slot read missed it, stranding the consumed pairing under the
old name. stageServerCredentialSlotMove (relocated to
device_credential_residue.go) now also moves the raw file slots via
os.Rename (never deviceSecretFileSet — its first current commit would
write the marker), with rollback covering both layers; regression
test plants the raw pending file. (2) command_server_test.go split:
remove coverage moved to command_server_remove_test.go.
